requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.6
django-cors-headers==3.7.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyotp==2.6.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
